requirements.txt
numpy
opencv-python
Pillow
pillow-heif